+++ lib/src/icons.dart
@@ -12,7 +12,7 @@ class GYWIcon {
   );
 
   /// Path of the associated file in the assets folder
-  String get path => "assets/icons/$filename.jpg";
+  String get path => "assets/icons/$filename.png";
 
   /// Width (in pixel) of the icon
   double get width => 80.0;
@@ -23,154 +23,32 @@ class GYWIcon {
 
 /// Icons supported by aRdent devices
 class GYWIcons {
-  // Icons from IconAwesome
-  static const GYWIcon arrowDown = GYWIcon(
-    "arrowDown",
-    "IconAwesomeArrowDown",
-  );
-  static const GYWIcon arrowLeft = GYWIcon(
-    "arrowLeft",
-    "IconAwesomeArrowLeft",
-  );
-  static const GYWIcon arrowRight = GYWIcon(
-    "arrowRight",
-    "IconAwesomeArrowRight",
-  );
-  static const GYWIcon arrowUp = GYWIcon(
-    "arrowUp",
-    "IconAwesomeArrowUp",
-  );
-  static const GYWIcon arrowAlt = GYWIcon(
-    "arrowAlt",
-    "IconAwesomeArrowsAlt",
-  );
-  static const GYWIcon arrowExpandAlt = GYWIcon(
-    "arrowExpandAlt",
-    "IconAwesomeExpandArrowsAlt",
-  );
-  static const GYWIcon arrowLocation = GYWIcon(
-    "arrowLocation",
-    "IconAwesomeLocationArrow",
-  );
-  static const GYWIcon batteryFull = GYWIcon(
-    "batteryFull",
-    "IconAwesomeBatteryFull",
-  );
-  static const GYWIcon batteryHalf = GYWIcon(
-    "batteryHalf",
-    "IconAwesomeBatteryHalf",
-  );
-  static const GYWIcon batteryQuarter = GYWIcon(
-    "batteryQuarter",
-    "IconAwesomeBatteryQuarter",
-  );
-  static const GYWIcon batteryEmpty = GYWIcon(
-    "batteryEmpty",
-    "IconAwesomeBatteryEmpty",
-  );
-  static const GYWIcon dropbox = GYWIcon(
-    "dropbox",
-    "IconAwesomeDropbox",
-  );
-  static const GYWIcon flag = GYWIcon(
-    "flag",
-    "IconAwesomeFlag",
-  );
-  static const GYWIcon flagCheckered = GYWIcon(
-    "flagCheckered",
-    "IconAwesomeFlagCheckered",
-  );
-
-  // Icons from IconFeather
-  static const GYWIcon message = GYWIcon(
-    "message",
-    "IconFeatherMessageCircle",
-  );
-
-  // Icons from IconIonicIos
-  static const GYWIcon speed = GYWIcon(
-    "speed",
-    "IconIonicIosSpeedometer",
-  );
-  static const GYWIcon send = GYWIcon(
-    "send",
-    "IconIonicIosSend",
-  );
-  static const GYWIcon notification = GYWIcon(
-    "notification",
-    "IconIonicIosNotifications",
-  );
-  static const GYWIcon information = GYWIcon(
-    "information",
-    "IconIonicIosInformationCircle",
-  );
-  static const GYWIcon checkbox = GYWIcon(
-    "checkbox",
-    "IconIonicIosCheckbox",
-  );
-
-  // Icons from IconMap
-  static const GYWIcon food = GYWIcon(
-    "food",
-    "IconMapFood",
-  );
-
-  // Icons from IconMaterial
-  static const GYWIcon mailbox = GYWIcon(
-    "mailbox",
-    "IconMaterialMarkunreadMailbox",
-  );
-  static const GYWIcon arrowUpLeft = GYWIcon(
-    "arrowUpLeft",
-    "IconMaterialSubdirectoryArrowLeft",
-  );
-  static const GYWIcon arrowUpRight = GYWIcon(
-    "arrowUpRight",
-    "IconMaterialSubdirectoryArrowRight",
-  );
-  static const GYWIcon arrowDownLeft = GYWIcon(
-    "arrowDownLeft",
-    "IconMaterialSubdirectoryArrowLeft1",
-  );
-  static const GYWIcon arrowDownRight = GYWIcon(
-    "arrowDownRight",
-    "IconMaterialSubdirectoryArrowRight1",
-  );
-
-  // Icons from IconMetro
-  static const GYWIcon fire = GYWIcon(
-    "fire",
-    "IconMetroFire",
-  );
+  // Icons from Google Icons
+  static const GYWIcon down = GYWIcon("Down", "down");
+  static const GYWIcon left = GYWIcon("Left", "left");
+  static const GYWIcon right = GYWIcon("Right", "right");
+  static const GYWIcon up = GYWIcon("Up", "up");
+  static const GYWIcon next = GYWIcon("Next", "next");
+  static const GYWIcon previous = GYWIcon("Previous", "prev");
+  static const GYWIcon checkbox = GYWIcon("Checkbox checked", "check");
+  static const GYWIcon checkboxEmpty = GYWIcon("Checkbox empty", "uncheck");
+  static const GYWIcon location = GYWIcon("Location", "location");
+  static const GYWIcon done = GYWIcon("Done", "done");
+  static const GYWIcon edit = GYWIcon("Edit", "edit");
+  static const GYWIcon rename = GYWIcon("Rename", "rename");
 
   /// List of all icons
   static const List<GYWIcon> values = [
-    arrowDown,
-    arrowLeft,
-    arrowRight,
-    arrowUp,
-    arrowAlt,
-    arrowExpandAlt,
-    arrowLocation,
-    batteryFull,
-    batteryHalf,
-    batteryQuarter,
-    batteryEmpty,
-    dropbox,
-    flag,
-    flagCheckered,
-    message,
-    speed,
-    send,
-    notification,
-    information,
+    down,
+    up,
+    right,
+    left,
+    next,
+    previous,
     checkbox,
-    food,
-    mailbox,
-    arrowUpLeft,
-    arrowUpRight,
-    arrowDownLeft,
-    arrowDownRight,
-    fire,
+    checkboxEmpty,
+    done,
+    edit,
+    rename,
   ];
 }