+++ test/drawing_test.dart
@@ -4,95 +4,6 @@ import 'package:flutter_gyw/flutter_gyw.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 main() {
-  group("Bluetooth JSON", () {
-    test('white screen', () {
-      Drawing drawing = const WhiteScreen();
-
-      final bluetoothJson = drawing.toBluetoothJson();
-      expect(bluetoothJson["type"], "white_screen");
-    });
-
-    test('Text', () {
-      Drawing drawing = const TextDrawing(text: "Text");
-
-      final bluetoothJson = drawing.toBluetoothJson();
-      expect(bluetoothJson["type"], "text");
-      expect(bluetoothJson["data"], "Text");
-    });
-
-    test('Text with position', () {
-      Drawing drawing = const TextDrawing(
-        text: "Text with position",
-        left: 100,
-        top: 200,
-      );
-
-      final bluetoothJson = drawing.toBluetoothJson();
-      expect(bluetoothJson["type"], "text");
-      expect(bluetoothJson["data"], "Text with position");
-      expect(bluetoothJson["x_start"], 100);
-      expect(bluetoothJson["y_start"], 200);
-    });
-
-    test('Text with font', () {
-      const font = GYWFont("Test", 42);
-
-      Drawing drawing = const TextDrawing(
-        text: "Text with font",
-        font: font,
-      );
-
-      final bluetoothJson = drawing.toBluetoothJson();
-      expect(bluetoothJson["type"], "text");
-      expect(bluetoothJson["data"], "Text with font");
-      expect(bluetoothJson["title"], 42);
-    });
-
-    test('Text with font and position', () {
-      const font = GYWFont("Test", 42);
-
-      Drawing drawing = const TextDrawing(
-        text: "Text with font and position",
-        left: 150,
-        top: 250,
-        font: font,
-      );
-
-      final bluetoothJson = drawing.toBluetoothJson();
-      expect(bluetoothJson["type"], "text");
-      expect(bluetoothJson["data"], "Text with font and position");
-      expect(bluetoothJson["x_start"], 150);
-      expect(bluetoothJson["y_start"], 250);
-      expect(bluetoothJson["title"], 42);
-    });
-
-    test('Icon', () {
-      const icon = GYWIcon("Icon", "icon");
-
-      Drawing drawing = const IconDrawing(icon);
-
-      final bluetoothJson = drawing.toBluetoothJson();
-      expect(bluetoothJson["type"], "memory");
-      expect(bluetoothJson["data"], "icon");
-    });
-
-    test('Icon with position', () {
-      const icon = GYWIcon("Icon with position", "icon_with_position");
-
-      Drawing drawing = const IconDrawing(
-        icon,
-        left: 120,
-        top: 220,
-      );
-
-      final bluetoothJson = drawing.toBluetoothJson();
-      expect(bluetoothJson["type"], "memory");
-      expect(bluetoothJson["data"], "icon_with_position");
-      expect(bluetoothJson["x_start"], 120);
-      expect(bluetoothJson["y_start"], 220);
-    });
-  });
-
   group("JSON", () {
     test('white screen', () {
       Drawing drawing = const WhiteScreen();
@@ -117,7 +28,11 @@ main() {
     });
 
     test('Text with font', () {
+<<<<<<< HEAD
       const font = GYWFonts.bigBasic;
+=======
+      const font = GYWFonts.medium;
+>>>>>>> master
 
       Drawing drawing = const TextDrawing(
         text: "Text with font",
@@ -128,7 +43,11 @@ main() {
     });
 
     test('Text with font and position', () {
+<<<<<<< HEAD
       const font = GYWFonts.bigBasic;
+=======
+      const font = GYWFonts.medium;
+>>>>>>> master
 
       Drawing drawing = const TextDrawing(
         text: "Text with font and position",
@@ -141,7 +60,11 @@ main() {
     });
 
     test('Icon', () {
+<<<<<<< HEAD
       const icon = GYWIcons.arrowAlt;
+=======
+      const icon = GYWIcons.checkbox;
+>>>>>>> master
 
       Drawing drawing = const IconDrawing(icon);
 
@@ -149,7 +72,11 @@ main() {
     });
 
     test('Icon with position', () {
+<<<<<<< HEAD
       const icon = GYWIcons.arrowUp;
+=======
+      const icon = GYWIcons.up;
+>>>>>>> master
 
       Drawing drawing = const IconDrawing(
         icon,